requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.0.1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.8
imageio==2.37.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.5.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
murmurhash==1.0.13
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
preshed==3.0.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.1.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rich-toolkit==0.14.8
rignore==0.6.4
rpds-py==0.20.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.5.1
scipy==1.14.1
Send2Trash==1.8.3
sentry-sdk==2.33.2
setuptools==74.0.0
shapely==2.1.1
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
terminado==0.18.1
thefuzz==0.22.1
thinc==8.3.6
threadpoolctl==3.5.0
tifffile==2025.6.11
tinycss2==1.3.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typer==0.16.0
types-python-dateutil==2.9.0.20240821
typing-inspection==0.4.1
typing_extensions==4.12.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.35.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
